refactor(api): move eventLogginJobRepository to shared
