+++ api/src/prescription/campaign-participation/domain/usecases/index.js
@@ -10,7 +10,6 @@ import * as badgeRepository from '../../../../evaluation/infrastructure/reposito
 import * as competenceEvaluationRepository from '../../../../evaluation/infrastructure/repositories/competence-evaluation-repository.js';
 import * as stageAcquisitionRepository from '../../../../evaluation/infrastructure/repositories/stage-acquisition-repository.js';
 import * as stageRepository from '../../../../evaluation/infrastructure/repositories/stage-repository.js';
-import { eventLoggingJobRepository } from '../../../../identity-access-management/infrastructure/repositories/jobs/event-logging-job.repository.js';
 import * as userRepository from '../../../../identity-access-management/infrastructure/repositories/user.repository.js';
 import { featureToggles } from '../../../../shared/infrastructure/feature-toggles/index.js';
 import * as areaRepository from '../../../../shared/infrastructure/repositories/area-repository.js';
@@ -18,6 +17,7 @@ import * as assessmentRepository from '../../../../shared/infrastructure/reposit
 import * as badgeForCalculationRepository from '../../../../shared/infrastructure/repositories/badge-for-calculation-repository.js';
 import * as competenceRepository from '../../../../shared/infrastructure/repositories/competence-repository.js';
 import { repositories as injectedSharedRepositories } from '../../../../shared/infrastructure/repositories/index.js';
+import { eventLoggingJobRepository } from '../../../../shared/infrastructure/repositories/jobs/event-logging-job.repository.js';
 import * as organizationLearnerRepository from '../../../../shared/infrastructure/repositories/organization-learner-repository.js';
 import * as organizationRepository from '../../../../shared/infrastructure/repositories/organization-repository.js';
 import { injectDependencies } from '../../../../shared/infrastructure/utils/dependency-injection.js';
